<ORDERS> | Added Kafka to monitor Order status | <Suhas>

+++ order/app/orderHandler.go
@@ -1,11 +1,15 @@
 package app
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/sirupsen/logrus"
+	pb "github.com/swiggy-2022-bootcamp/cdp-team4/order/app/protobuf"
 	"github.com/swiggy-2022-bootcamp/cdp-team4/order/domain"
+	"google.golang.org/grpc"
 )
 
 type OrderHandler struct {
@@ -20,9 +24,19 @@ type ProductRecordDTO struct {
 
 type OrderRecordDTO struct {
 	UserID    string             `json:"user_id"`
+	OrderID   string             `json:"order_id"`
 	Status    string             `json:"status"`
 	Products  []ProductRecordDTO `json:"products"`
 	TotalCost int16              `json:"total_cost"`
+}
+
+type OrderConfirmResponseDTO struct {
+	UserID                string `json:"user_id"`
+	OrderID               string `json:"order_id"`
+	Status                string `json:"status"`
+	TotalCost             int16  `json:"total_cost"`
+	ShippingPrice         int16  `json:"shipping_price"`
+	RewardspointsConsumed int16  `json:"reward_points"`
 }
 
 // Create Order
@@ -39,6 +53,8 @@ func (oh OrderHandler) handleOrder() gin.HandlerFunc {
 		var orderDto OrderRecordDTO
 		if err := ctx.BindJSON(&orderDto); err != nil {
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadRequest}).
+				Error("Error while creating order")
 			return
 		}
 
@@ -52,7 +68,9 @@ func (oh OrderHandler) handleOrder() gin.HandlerFunc {
 			int(orderDto.TotalCost),
 		)
 		if err != nil {
-			ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadRequest}).
+				Error("Error while creating order")
+			ctx.JSON(http.StatusBadGateway, gin.H{"message": err.Error()})
 			return
 		}
 		ctx.JSON(http.StatusAccepted, gin.H{"message": "Order Record Added", "order id": res})
@@ -70,10 +88,18 @@ func (oh OrderHandler) handleOrder() gin.HandlerFunc {
 func (oh OrderHandler) HandleGetOrderRecordByID() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		id := ctx.Param("id")
+		if id == "" {
+			log.WithFields(logrus.Fields{"message": "Invalid ID", "status": http.StatusBadRequest}).
+				Error("Error while Getting order by order id")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Invalid ID"})
+			return
+		}
 		fmt.Println(id)
 		res, err := oh.OrderService.GetOrderById(id)
 
 		if err != nil {
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadRequest}).
+				Error("Record not found")
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Record not found"})
 			return
 		}
@@ -93,10 +119,18 @@ func (oh OrderHandler) HandleGetOrderRecordByID() gin.HandlerFunc {
 func (oh OrderHandler) HandleGetOrderRecordsByUserID() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		id := ctx.Param("user_id")
+		if id == "" {
+			log.WithFields(logrus.Fields{"message": "Invalid ID", "status": http.StatusBadRequest}).
+				Error("Error while getting order by user id")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Invalid ID"})
+			return
+		}
 		fmt.Println(id)
 		record, err := oh.OrderService.GetOrderByUserId(id)
 
 		if err != nil {
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadRequest}).
+				Error("Record not found")
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Record not found"})
 			return
 		}
@@ -123,9 +157,17 @@ func (oh OrderHandler) HandleGetOrderRecordsByUserID() gin.HandlerFunc {
 func (oh OrderHandler) HandleGetOrderRecordsByStatus() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		status := ctx.Param("status")
+		if status == "" {
+			log.WithFields(logrus.Fields{"message": "No Status param", "status": http.StatusBadRequest}).
+				Error("Record not found")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "No Status param"})
+			return
+		}
 		record, err := oh.OrderService.GetOrderByStatus(status)
 
 		if err != nil {
+			log.WithFields(logrus.Fields{"message": "Record not found", "status": http.StatusBadRequest}).
+				Error("Record not found")
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Record not found"})
 			return
 		}
@@ -153,6 +195,8 @@ func (oh OrderHandler) HandleGetAllRecords() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		record, err := oh.OrderService.GetAllOrders()
 		if err != nil {
+			log.WithFields(logrus.Fields{"message": "Record not found", "status": http.StatusBadRequest}).
+				Error("Record not found")
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Record not found"})
 			return
 		}
@@ -182,12 +226,16 @@ func (oh OrderHandler) handleUpdateOrderStatus() gin.HandlerFunc {
 			Status string `json:"status"`
 		}
 		if err := ctx.BindJSON(&requestDTO); err != nil {
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadRequest}).
+				Error(err.Error())
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
 			return
 		}
 
 		ok, err := oh.OrderService.UpdateOrderStatus(requestDTO.Id, requestDTO.Status)
 		if !ok {
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadRequest}).
+				Error(err.Error())
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
 			return
 		}
@@ -206,15 +254,84 @@ func (oh OrderHandler) handleUpdateOrderStatus() gin.HandlerFunc {
 func (oh OrderHandler) HandleDeleteOrderById() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		id := ctx.Param("id")
+		if id == "" {
+			log.WithFields(logrus.Fields{"message": "Invalid ID", "status": http.StatusBadRequest}).
+				Error("Invalid ID")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Invalid ID"})
+			return
+		}
 		fmt.Println(id)
 		_, err := oh.OrderService.DeleteOrderById(id)
 
 		if err != nil {
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadGateway}).
+				Error(err.Error())
 			ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
 			return
 		}
 
 		ctx.JSON(http.StatusAccepted, gin.H{"message": "Deleted Succesfully"})
+	}
+}
+
+func (oh OrderHandler) HandleAddOrderFromCart() gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		id := ctx.Param("user_id")
+		if id == "" {
+			log.WithFields(logrus.Fields{"message": "Invalid ID", "status": http.StatusBadRequest}).
+				Error("Invalid ID")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Invalid ID"})
+			return
+		}
+		conn, err := grpc.Dial("localhost:7899", grpc.WithInsecure())
+		if err != nil {
+			log.WithFields(logrus.Fields{"message": "Error while making connection", "status": http.StatusBadGateway}).
+				Error("Error while making connection")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Error while making connection"})
+			return
+		}
+
+		// Create a client instance
+		c := pb.NewCheckoutClient(conn)
+		resp, err := c.OrderOverview(context.Background(), &pb.OverviewRequest{
+			UserID: id,
+		})
+		if err != nil {
+			log.WithFields(logrus.Fields{"message": "Error while getting grpc response", "status": http.StatusBadGateway}).
+				Error("Error while getting grpc response")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Error while getting grpc response"})
+			return
+		}
+		var newpdto []ProductRecordDTO
+		for _, prod := range resp.Item {
+			newpdto = append(newpdto, ProductRecordDTO{
+				Product:  prod.Name,
+				Quantity: int(prod.Qty),
+				Cost:     int16(prod.Price),
+			})
+		}
+		product_quantity, product_cost := convertProductsDTOtoMaps(newpdto)
+		res, result := oh.OrderService.CreateOrder(
+			resp.UserID,
+			"Pending",
+			product_quantity,
+			product_cost,
+			int(resp.TotalPrice),
+		)
+		if result != nil {
+			log.WithFields(logrus.Fields{"message": err.Error(), "status": http.StatusBadGateway}).
+				Error("Error while creating order")
+			ctx.JSON(http.StatusBadRequest, gin.H{"message": "Error while creating order"})
+			return
+		}
+		ctx.JSON(http.StatusAccepted, gin.H{"record": OrderConfirmResponseDTO{
+			UserID:                resp.UserID,
+			OrderID:               res,
+			Status:                "Pending",
+			TotalCost:             int16(resp.TotalPrice),
+			ShippingPrice:         int16(resp.ShippingPrice),
+			RewardspointsConsumed: int16(resp.RewardPointsConsumed),
+		}})
 	}
 }
 
@@ -237,6 +354,7 @@ func convertOrderModeltoOrderDTO(order domain.Order) OrderRecordDTO {
 		UserID:    order.UserID,
 		Status:    order.Status,
 		Products:  prodcts,
+		OrderID:   order.ID,
 		TotalCost: int16(order.TotalCost),
 	}
 }
